Refactor for adding ships to a grid
