+++ test/battleship/game/grid_test.go
@@ -1,0 +1,17 @@
+package game
+
+import (
+	"battleship/game"
+	"testing"
+
+	"github.com/corbym/gocrest/is"
+	"github.com/corbym/gocrest/then"
+)
+
+func TestAddShip(t *testing.T) {
+	grid, _ := game.NewGrid(3)
+
+	grid = game.AddShip(grid, game.Ship{1})
+
+	then.AssertThat(t, len(grid.Ships), is.EqualTo(1).Reason("Added one ship"))
+}
